third_party/WebKit/LayoutTests/inspector/console/console-error-on-call-frame-expected.txt

CONSOLE MESSAGE: line 2: 42
Tests that console.error does not throw exception when executed in console on call frame.

Debugger was enabled.
Set timer for test function.
Script execution paused.
Script execution resumed.
setTimeout(testFunction, 0)
1
console.error(42)
42testFunction console-error-on-call-frame.html:10
undefined
Debugger was disabled.
